@@ -3,7 +3,7 @@ description: Internal data package of WebTrit Phone application.
 version: 1.0.0
 
 environment:
-  sdk: '>=2.18.6 <3.0.0'
+  sdk: '>=2.19.0 <3.0.0'
 
 dependencies:
   clock: ^1.1.1
