@@ -3,7 +3,7 @@ description: WebTrit Signalling client.
 version: 1.0.0
 
 environment:
-  sdk: '>=2.17.0 <3.0.0'
+  sdk: '>=2.19.0 <3.0.0'
 
 dependencies:
   equatable: ^2.0.5
