@@ -4,8 +4,8 @@ publish_to: 'none'
 version: 1.1.0+1
 
 environment:
-  sdk: ">=2.18.0 <3.0.0"
-  flutter: ^3.3.0
+  sdk: ">=2.19.0 <3.0.0"
+  flutter: ^3.7.0
 
 dependencies:
   flutter:
