@@ -37,7 +37,7 @@ class RecentScaffold extends StatelessWidget {
                 ),
                 Text(
                   recent.name,
-                  style: themeData.textTheme.headline4,
+                  style: themeData.textTheme.headlineMedium,
                   textAlign: TextAlign.center,
                 ),
                 const Divider(
