@@ -5,11 +5,8 @@ import 'package:flutter/material.dart';
 class ProgressOverlay extends StatelessWidget {
   static void insert<T>(BuildContext context, Future<T> removeFuture) {
     final overlayEntry = OverlayEntry(builder: (context) => const ProgressOverlay._());
-    final overlay = Overlay.of(context);
-    if (overlay != null) {
-      overlay.insert(overlayEntry);
-      removeFuture.whenComplete(() => overlayEntry.remove()).ignore();
-    }
+    Overlay.of(context).insert(overlayEntry);
+    removeFuture.whenComplete(() => overlayEntry.remove()).ignore();
   }
 
   const ProgressOverlay._({
